bugfix: Kill the process group, not just the parent process (#328), thanks @ioqy, @vvrein
